ChannelServer can now authenticate to the world server and listen to the port it's asked to
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@230 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/WorldServer/WorldServerAcceptPlayerPacket.h
@@ -21,7 +21,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 class WorldServerAcceptPlayer;
 
 namespace WorldServerAcceptPlayerPacket {
-	void connect(WorldServerAcceptPlayer *player, short channel, int port);
+	void connect(WorldServerAcceptPlayer *player, int channel, int port);
 };
 
 #endif
